cleaning up mock automaton
