+++ consensus/src/authority/mod.rs
@@ -170,7 +170,6 @@ mod tests {
     use engine::Engine;
     use futures::{channel::mpsc, StreamExt};
     use governor::Quota;
-    use mocks::{Application, Config as ApplicationConfig, Progress};
     use prometheus_client::registry::Registry;
     use prover::Prover;
     use std::{
@@ -179,156 +178,6 @@ mod tests {
         sync::{Arc, Mutex},
         time::Duration,
     };
-
-    type HeightActivity = HashMap<Height, HashMap<Digest, HashSet<PublicKey>>>;
-    type Faults = HashMap<PublicKey, HashMap<View, HashSet<Activity>>>;
-
-    #[derive(Clone)]
-    struct TestSupervisor<C: Scheme, H: Hasher> {
-        participants: BTreeMap<View, (HashSet<PublicKey>, Vec<PublicKey>)>,
-
-        prover: Prover<C, H>,
-
-        proposals: Arc<Mutex<HeightActivity>>,
-        votes: Arc<Mutex<HeightActivity>>,
-        finalizes: Arc<Mutex<HeightActivity>>,
-        faults: Arc<Mutex<Faults>>,
-    }
-
-    impl<C: Scheme, H: Hasher> TestSupervisor<C, H> {
-        fn new(prover: Prover<C, H>, participants: BTreeMap<View, Vec<PublicKey>>) -> Self {
-            let mut parsed_participants = BTreeMap::new();
-            for (view, mut validators) in participants.into_iter() {
-                let mut set = HashSet::new();
-                for validator in validators.iter() {
-                    set.insert(validator.clone());
-                }
-                validators.sort();
-                parsed_participants.insert(view, (set.clone(), validators));
-            }
-            Self {
-                participants: parsed_participants,
-                prover,
-                proposals: Arc::new(Mutex::new(HashMap::new())),
-                votes: Arc::new(Mutex::new(HashMap::new())),
-                finalizes: Arc::new(Mutex::new(HashMap::new())),
-                faults: Arc::new(Mutex::new(HashMap::new())),
-            }
-        }
-    }
-
-    impl<C: Scheme, H: Hasher> Supervisor for TestSupervisor<C, H> {
-        type Index = View;
-        type Seed = ();
-
-        fn leader(&self, index: Self::Index, _seed: Self::Seed) -> Option<PublicKey> {
-            let closest = match self.participants.range(..=index).next_back() {
-                Some((_, p)) => p,
-                None => {
-                    panic!("no participants in required range");
-                }
-            };
-            Some(closest.1[index as usize % closest.1.len()].clone())
-        }
-
-        fn participants(&self, index: Self::Index) -> Option<&Vec<PublicKey>> {
-            let closest = match self.participants.range(..=index).next_back() {
-                Some((_, p)) => p,
-                None => {
-                    panic!("no participants in required range");
-                }
-            };
-            Some(&closest.1)
-        }
-
-        fn is_participant(&self, index: Self::Index, candidate: &PublicKey) -> Option<bool> {
-            let closest = match self.participants.range(..=index).next_back() {
-                Some((_, p)) => p,
-                None => {
-                    panic!("no participants in required range");
-                }
-            };
-            Some(closest.0.contains(candidate))
-        }
-
-        async fn report(&mut self, activity: Activity, proof: Proof) {
-            // We check signatures for all messages to ensure that the prover is working correctly
-            // but in production this isn't necessary (as signatures are already verified in
-            // consensus).
-            match activity {
-                NOTARIZE => {
-                    // TODO: use payload digest?
-                    let (index, _, payload, public_key) =
-                        self.prover.deserialize_notarize(proof, true).unwrap();
-                    self.votes
-                        .lock()
-                        .unwrap()
-                        .entry(index.height)
-                        .or_default()
-                        .entry(payload)
-                        .or_default()
-                        .insert(public_key);
-                }
-                FINALIZE => {
-                    let (index, _, payload, public_key) =
-                        self.prover.deserialize_finalize(proof, true).unwrap();
-                    self.finalizes
-                        .lock()
-                        .unwrap()
-                        .entry(index.height)
-                        .or_default()
-                        .entry(payload)
-                        .or_default()
-                        .insert(public_key);
-                }
-                CONFLICTING_NOTARIZE => {
-                    let (public_key, view) = self
-                        .prover
-                        .deserialize_conflicting_notarize(proof, true)
-                        .unwrap();
-                    self.faults
-                        .lock()
-                        .unwrap()
-                        .entry(public_key)
-                        .or_default()
-                        .entry(view)
-                        .or_default()
-                        .insert(CONFLICTING_NOTARIZE);
-                }
-                CONFLICTING_FINALIZE => {
-                    let (public_key, view) = self
-                        .prover
-                        .deserialize_conflicting_finalize(proof, true)
-                        .unwrap();
-                    self.faults
-                        .lock()
-                        .unwrap()
-                        .entry(public_key)
-                        .or_default()
-                        .entry(view)
-                        .or_default()
-                        .insert(CONFLICTING_FINALIZE);
-                }
-                NULLIFY_AND_FINALIZE => {
-                    let (public_key, view) = self
-                        .prover
-                        .deserialize_nullify_finalize(proof, true)
-                        .unwrap();
-                    self.faults
-                        .lock()
-                        .unwrap()
-                        .entry(public_key)
-                        .or_default()
-                        .entry(view)
-                        .or_default()
-                        .insert(NULLIFY_AND_FINALIZE);
-                }
-                a => {
-                    panic!("unexpected activity: {}", a);
-                }
-            }
-        }
-    }
 
     #[test_traced]
     fn test_all_online() {
